Bump versions of Node.js, npm and MongoDB for the Windows `dev_bundle`.
In a similar fashion as:

* f51c6822ae13a6c5d90f4ca02eeceb91d51a067a (Node.js)
* 8ea41588ea535bb68d04b9beb7d9229cd9008c61 (npm)
* 3470516f9324e6f16e037b3c774ebcc37edff143 (MongoDB)

I overlooked the multiplicity of these version variables when updating the major package versions during the release of Meteor 1.4.3.2.

While I intend on never making this mistake again, I've also opened meteor/meteor#8505 which consolidates these variables to a single location (in `build-dev-bundle-common.sh`) to hopefully prevent this from happening to anyone else!