Use UglifyJS when possible, but fall back to Babel.minify.
This gives us all the benefits of @sethmurphy18's great work without the
additional performance hit of using Babili for everything.

With this commit:

  % meteor create min-test
  % cd min-test
  % time meteor build ../min-build
  meteor build ../min-build  39.26s user 5.49s system 112% cpu 39.647 total

Without this commit:

  meteor build ../min-build  59.28s user 5.23s system 107% cpu 1:00.08 total

Keep in mind this is the best-case scenario for Babili (a very small app).
Larger apps take much longer for Babili to minify.

cc @abernix @glasser @sethmurphy18