[backport] Explicitly join the fields array to a comma-delimited string.
This is a backport of 2c5dda1675a9e3526e3a926d3ec2b45d0fab14c6gu for the
Meteor 1.4.2.x series.

With https://github.com/meteor/meteor/pull/8261 merged, arrays are no longer concatenated (with commas) into a string during `HTTP.call` requests so we need to specify the exact behavior we want. Any previous behavior of this type was purely accidental.

The definition of this field in Facebook's API is that it should be a comma-separated field:

https://developers.facebook.com/docs/graph-api/using-graph-api