Add TOOL_NODE_FLAGS support for Windows in the same way as Unix. (#8517)
In Unix we support adding `TOOL_NODE_FLAGS` when debugging the Meteor
tool itself, however Windows did not currently support it.  This should
add that support.  As this is a .bat file, I have every reason to
believe that this syntax should work on older versions of Windows as the
syntax of .bat files hasn't changed much.

Helps with meteor/meteor#8513.