Add message about `MAIL_URL` scheme in `History.md`.
This should act as a reminder for those who may be using `smtp://` when
they should be using `smtps://`.  Previously, the scheme was implied by
the port number, however this is not the ideal behavior in the same way
that it is unsafe to assume that port 443 is always `https://`.