In case of double-wrapping Promise#then, disable previous wrappers.
Although the current version of the meteor-promise package is careful not
to double-wrap the Promise.prototype.then method if .makeCompatible is
called multiple times, it's possible we may encounter a .then method
defined by an older version of this package.

We would ideally like to remove/skip the older wrapper, so that it won't
double-wrap the onRe{solv,ject}ed callbacks, but we don't have direct
access to the original es6PromiseThen function that should be called
instead, so the best we can do is to trick the older wrapper into thinking
that Promise.Fiber is not a function, so that it will simply call its own
version of es6PromiseThen with no additional wrapping of callbacks.