Bump package versions for 1.5-rc.3 release.
Something went wrong with 1.5-rc.2 such that publishing meteor-tool failed
with the following error on all platforms:

  While publishing package build for meteor-tool:
  error: Cannot override existing build

This makes no sense to me, but it necessitates another RC release. To be
clear, meteor-tool@1.5.0-rc.2 has not been successfully published on any
platform, so please don't bother trying to update to it.