Change default `from` email for `accounts-password` to an "example".
Currently, the default "from" address for `meteor-accounts` is
"no-reply@meteor.com".  While this works for many users, and granted it
is a "no reply" address and there should be no expectation of the
address working, it contributes to a negative spam rating for the
"meteor.com" domain and a surplus of extra e-mail.

The correct way to set the default "from" address is by setting:

    Accounts.emailTemplates.from = "Name <email@domain.com>";

As per the documentation here:

    https://docs.meteor.com/api/passwords.html#Accounts-emailTemplates

By changing it to "example.com", and making a more obvious "example"
out of the name ("Accounts Example"), it should encourage users to
actually change the address to something more reasonable especially
since many e-mail providers will also reject mail coming from
"example.com", which should provide a clear warning to those who have
their e-mail misconfigured.