Update rimraf to latest version to hopefully help ENOTEMPTY errors.
There have been a number of commits in the history for `rimraf` which
indicate others have also been struggling with ENOTEMPTY, specifically
on Windows.

https://github.com/isaacs/rimraf/commits/master

This commit, included in 2.6.1 takes a relatively aggressive approach:

https://github.com/isaacs/rimraf/commit/d53235de863cfb00f691456288fcf3abbced7f6c

However, due to the way the Windows filesystem is designed, this
approach may be the only way of coping with file-handles which are
cached by the OS and not released immediately upon closing a file
(in other words, at a time that Meteor would think it'd be safe to
remove the file/directory).

Attempts to help with meteor/meteor#8485.