Append `_1` to `coffeescript` npm-wrapper package.
As discussed in https://github.com/meteor/meteor/pull/8777 it seems best
to always have the suffix on npm-wrapper packages.

If anything, as a reminder when bumping the version on wrapper packages,
but also just to generally make it more clear if a version suffixed with
`_1` is actually higher than a non-suffixed version or not.