Don't put `node-gyp` into a tiered `node_modules` on Windows dev bundle.
This was preventing `node-gyp` from installing the Node header files on
Windows and was the reason that `minimatch` was not being found, as seen
in https://github.com/meteor/meteor/pull/8831.

The `minimatch` module was present, but it was just in `dev_bundle/lib`,
not in `dev_bundle/lib/node_modules/npm/node_modules`.

This expecation may have been expected from older versions of npm but is
no longer the case.  This replicates the behavior of the Unix
`generate_dev_bundle.sh` script, which also does not nest `node-gyp`.

/cc @benjamn