npm-mongo: Upgrade to npm mongo@2.2.29 to fix arbiter login attempts
Meteor 1.4.3.2 upgraded npm mongo to 2.2.24 in order to fix an issue
with MongoDB arbiter login attempts, see Meteor issue 8449 and Meteor
1.4.3.2 changelog.

But the MongoDB upstream fix was incomplete.

This commits upgrades npm mongo further to 2.2.29 in order to get the
fix for the related arbiter login issue NODE-981 on the MongoDB Jira,
which requires `mongo-core@>=2.1.10`, which requires `mongodb@>=2.2.26`;
see references below.

References:

 - Meteor issue 8449: <https://github.com/meteor/meteor/issues/8449>.
 - Meteor 1.4.3.2: <https://github.com/meteor/meteor/blob/devel/History.md#v1432-2017-03-14>
 - NODE-981 in MongoDB Jira: <https://jira.mongodb.org/browse/NODE-981>
 - mongodb-core 2.1.10 changelog: <https://github.com/christkv/mongodb-core/blob/2.0/HISTORY.md#2110-2017-04-18>,
 - mongodb 2.2.26 changelog: <https://github.com/mongodb/node-mongodb-native/blob/2.2/HISTORY.md#2226-2017-04-18>.

Signed-off-by: Steffen Prohaska <prohaska@zib.de>