Switch to using CircleCI 2.0 and various other test improvements.
This switches Meteor's CircleCI builds from Circle 1.0 to Circle 2.0 which
has a bit more control over the workflow.

Currently, this eliminates the existing ci.sh script which was already a bit
incompatible when I was attempting to run Windows builds on another environment.
It's possible that we should change this to a Node.js wrapper script.

Other improvements:

- We now store Core Dumps in build artifacts.  CircleCI 2.0 advertised this as
  one of the features of CircleCI 2.0, but honestly, it was far from
  straightforward.  Perhaps if we were using another Dockerimage, but it was
  far from as easy as flipping a switch.  In addition to saving the Core Dump,
  this also saves the Node.js binary which was included in the Dev Bundle.  This
  can be very handy for post-mortem debugging with tools like lldb, gdb, or mdb.

- Memory usage is now logged throughout the build via a background process which
  logs `ps` output to a file which is persisted to the build artifacts.  This
  should help identify if builds are terminating for some environmental reason.