Use Babel options appropriate for Node 8 to compile meteor/tools.
Related to https://github.com/meteor/meteor/pull/8728#issue-232369984,
though these changes apply only to the meteor/tools codebase. We still
need to make similar changes to the babel-compiler package so that
application code will benefit.

The assumption that we're using Node 8 also allows some simplifications to
the runtime polyfills that we use.