Fix html-scanner tests by never falsifying noLineNumbers when true.
We should never be generating line number comments for server code, since
they can cause problems with multi-line strings, and no one will see those
comments on the server anyway.

The html-scanner tests were breaking because we no longer compile
multi-line template strings for Node 8, yet we were annotating every line
with a line number comment (because we had a source map), and lines within
multi-line template strings were getting comments appended to them, which
thus became an observable part of the string.

On the client, we still compile template strings to normal string
literals, so the line number annotations are not a problem (for now).