Adjust the previous watcher check to handle moved files.
When checking the `entriesByIno` Map to see if an `entry` already exists for
the specified inode, also check to make sure the found `entry` is only
re-used if the current file watcher path matches the returned path. This
makes sure new file watchers are created for moved files (so files with the
same inode), instead of attempting to re-use a file watcher that's watching
an invalid path.