Support @~ version constraints and use them for core packages.
As discovered in #8947, creating an app with

  meteor create --release 1.4.4.1 new-app

does not actually constrain core packages to have the same minor versions
they had when Meteor 1.4.4.1 was published. Rather (as of Meteor 1.4.3),
the Meteor release only constrains core packages to have versions at least
as recent as the versions they had when the release was published, and to
have the same major version. Any later minor version is fair game.

What we need is a new kind of package version constraint, more like
semver's ~ range operator:
https://www.npmjs.com/package/semver#tilde-ranges-123-12-1

This commit implements that functionality, though it stops short of
supporting all possible semver range syntax.