Revert "Add mongo-dev-server package (#8853)"
This reverts commit 4d37a05fb33576b8f167168f4bc1b12821354615.

After git bisecting between origin/release-1.5 and origin/release-1.5.2, I
identified this commit as the culprit in recent failures of the modules
test app: https://circleci.com/gh/meteor/meteor/4857#tests/containers/3

Note that the modules test app seems to be failing only on Linux, and it
does pass reliably with this commit reverted. It must have something to do
with Mongo failing to start, and thus the "App running at" message never
appears, but I don't have a good theory why that might be.

The command to run just the modules test app is

  meteor self-test --history 1000 'modules - test app'

@zimme @hwillson @abernix any ideas?