Replace for-in loops in minimongo package with Object.keys/forEach.
This fixes a problem reported by @arggh:
https://github.com/meteor/meteor/pull/8913#issuecomment-322048927

The problem appears to have been introduced by this commit:
2c5094fa0cdd7b7ae649a2d843cb73888eff0c4b

The reason this commit was problematic is that it turned several
Object.keys/forEach loops into for-in loops, which caused a few unchanged
return statements to take on a new meaning: instead of "continuing" the
loop by returning from the forEach callback function, the return
statements now returned from the enclosing function:
https://github.com/meteor/meteor/commit/2c5094fa0cdd7b7ae649a2d843cb73888eff0c4b#commitcomment-23638205
https://github.com/meteor/meteor/commit/2c5094fa0cdd7b7ae649a2d843cb73888eff0c4b#commitcomment-23638216

Besides this refactoring hazard, what else is wrong with for-in loops?

First, for-in loops iterate over not only own but also inherited keys,
which makes them not exactly equivalent to Object.keys/forEach, which
iterates over only own keys. Another way to address that problem is to
initialize this.queries to Object.create(null) instead of {} (which I've
also done in this commit), so that there is no prototype chain.

Second, for-in loops are a trigger for deoptimization in V8:
https://github.com/petkaantonov/bluebird/wiki/Optimization-killers#5-for-in

Since 2c5094fa0cdd7b7ae649a2d843cb73888eff0c4b was supposed to be about
improving performance, and these changes introduced some logical problems
as well, I think it's best that we avoid for-in loops.