Build Node.js with ICU support.
To remain the same as official Node.js releases, we need to build with
the `small-icu` ICU (International Components for Unicode) package.
For Node.js 4.x this means ICU 56.x.  As found with `process.config` on
an official Node.js release.

See https://github.com/nodejs/node/wiki/Intl#configure-node-with-specific-icu-source.