When building Node.js, use `DESTDIR` to specify install location.
In an effort to maintain parity with the official Node.js build, which
sets `node_prefix` to `/`, this takes advantage of the `DESTDIR`
variable which changes the actual installation location used by
`make install`.  Without this, it relies on the `PREFIX` which is now
explicitly set to `/`, even though it won't be installed there.

See the `install.py` script for more information, in the form of code:
https://github.com/nodejs/node/blob/2a97eb61c65e23de3cee42c7b70f845b0c3a3d09/tools/install.py#L13

/cc @benjamn