Adjust force-ssl localhost regex to handle IPv4 in IPv6 addresses
`force-ssl`'s current `isLocalConnection` regex does not handle
IPv4-mapped and/or IPv4-compatible IPv6 addresses, when checking
if the connection is local. This means a local address of
`::ffff:127.0.0.1` is being marked as non-local. These changes
adjust the `isLocalConnection` regex to handle a wider range of
localhost address representations.

Fixes #9072.