Stop using path.join to require fiber/futures.
This misunderstanding of CommonJS module identifiers goes back to 2012:
https://github.com/meteor/meteor/commit/8f83b2c32e5e0b8fe1d89e26983799b8517cdc41

In Node, CommonJS module identifiers are always delimited by forward
slashes, even on Windows. Using path.join to abstract away the choice of
platform-specific delimiter is therefore pointless.

While this usage of path.join was always pointless, it was also mostly
harmless until #9095, when we stopped falling back to Node's native
`require` function to handle e.g. `Npm.require("fibers\\future")`. That PR
made the misuse problematic on Windows, so this commit fixes that.