Beta versions of coffeescript packages compatible with 1.6-beta.30.
New coffeescript* versions are necessary because we've bumped the major
versions of babel-compiler and babel-runtime.

To verify that these new versions work, create a new app with

  meteor create --release 1.6-beta.30 new-app

and then add coffeescript to it:

  cd new-app
  meteor add coffeescript@1.12.7-2-beta.30

If everything works, you should get -beta.30 versions of coffeescript and
coffeescript-compiler, instead of something earlier.

cc @GeoffreyBooth