Add DDP.onReconnect(), deprecate conn.onReconnect  (#9092)
* Fix bug #5665: add DDP.onReconnect(), deprecate conn.onReconnect

Deprecate use of `connection.onReconnect = func`. Instead, a new
`DDP.onReconnect(callback)` method should be used to register callbacks to call
when a connection reconnects. The connection that is reconnecting is passed as
the only argument to `callback`. This is used by the accounts system to relogin
on reconnects without interfering with other code which uses
`connection.onReconnect`.

* Adjust History entry, package versions, code cleanup