Don't strip the top directory when extracting our Node tarball.
Thanks to changes in the `build-node-for-dev-bundle.sh` script done in
0583e5883c611424076811441634b45c69062d07, we now build a tarball which
is identical to the structure provided by Node.js themselves.

While generally we are using the main Node releases, this will allow
our users to (if need be), use a tarball directly in place of their own
in production without additional changes.  Similarly, the only change
we need to make now when building the dev bundle is to use a different
URL.